deps

+++ ctxcdb/pebble/pebble.go
@@ -158,7 +158,7 @@ func New(file string, cache int, handles int, namespace string, readonly bool, e
 
 		// The size of memory table(as well as the write buffer).
 		// Note, there may have more than two memory tables in the system.
-		MemTableSize: memTableSize,
+		MemTableSize: uint64(memTableSize),
 
 		// MemTableStopWritesThreshold places a hard limit on the size
 		// of the existent MemTables(including the frozen one).
